quality: solve all gcc compiler warnings
- Solve all implicit numeric conversion warnings
- Solve all useless casting warnings
- Solve all implicit this capture warnings which deprecated in c++20
- Solve all unused variables warnings
- Solve multiline comment warning

Ticket: SOUR-76

+++ src/lib_gui/qt/element/QtTable.cpp
@@ -50,11 +50,11 @@ QtTable::QtTable(QWidget* parent) : QTableView(parent), m_rowsToFill(0) {
 QtTable::~QtTable() {}
 
 void QtTable::updateRows() {
-  while(model()->rowCount() <= m_rowsToFill) {
+  while(static_cast<float>(model()->rowCount()) <= m_rowsToFill) {
     model()->insertRow(model()->rowCount());
   }
 
-  while(model()->rowCount() > m_rowsToFill + 1) {
+  while(static_cast<float>(model()->rowCount()) > m_rowsToFill + 1) {
     int row = model()->rowCount() - 1;
     if(model()->index(row, 0).data(Qt::DisplayRole).toString().isEmpty()) {
       model()->removeRow(row);
@@ -63,7 +63,8 @@ void QtTable::updateRows() {
     }
   }
 
-  const int rowCount = model()->rowCount() > m_rowsToFill ? model()->rowCount() : static_cast<int>(m_rowsToFill);
+  const int rowCount = static_cast<float>(model()->rowCount()) > m_rowsToFill ? model()->rowCount() :
+                                                                                static_cast<int>(m_rowsToFill);
   const int width = static_cast<int>(IApplicationSettings::getInstanceRaw()->getFontSize() * 0.7 * int(1 + std::log10(rowCount)));
 
   verticalHeader()->setStyleSheet("::section { width: " + QString::number(width) + "px; }");
@@ -91,7 +92,7 @@ void QtTable::showFirstRow() {
 }
 
 void QtTable::showLastRow() {
-  if(m_rowsToFill <= getFilledRowCount()) {
+  if(m_rowsToFill <= static_cast<float>(getFilledRowCount())) {
     verticalScrollBar()->setValue(verticalScrollBar()->maximum());
   }
 }
