+++ src/lib_gui/qt/utility/QtScrollSpeedChangeListener.cpp
@@ -24,6 +24,6 @@ void QtScrollSpeedChangeListener::handleMessage(MessageScrollSpeedChange* messag
 
 void QtScrollSpeedChangeListener::doChangeScrollSpeed(float scrollSpeed) {
   if(m_scrollBar) {
-    m_scrollBar->setSingleStep(static_cast<int>(std::ceil(m_singleStep * scrollSpeed)));
+    m_scrollBar->setSingleStep(static_cast<int>(std::ceil(static_cast<float>(m_singleStep) * scrollSpeed)));
   }
 }
